+++ source/renderer/app/containers/settings/categories/GeneralSettingsPage.tsx
@@ -1,7 +1,6 @@
 import React, { Component } from 'react';
 import { observer, inject } from 'mobx-react';
 import GeneralSettings from '../../../components/settings/categories/GeneralSettings';
-import { rebuildApplicationMenu } from '../../../ipc/rebuild-application-menu';
 import type { InjectedProps } from '../../../types/injectedPropsType';
 
 @inject('stores', 'actions')
@@ -12,19 +11,12 @@ class GeneralSettingsPage extends Component<InjectedProps> {
     stores: null,
   };
   handleSelectItem = async (param: string, value: string) => {
-    const { actions, stores } = this.props;
-    const { areTermsOfUseAccepted: isNavigationEnabled } = stores.profile;
+    const { actions } = this.props;
     const { updateUserLocalSetting } = actions.profile;
     updateUserLocalSetting.trigger({
       param,
       value,
     });
-
-    if (param === 'locale') {
-      await rebuildApplicationMenu.send({
-        isNavigationEnabled,
-      });
-    }
   };
 
   render() {