+++ source/renderer/app/ipc/rebuild-application-menu.ts
@@ -5,7 +5,10 @@ import type {
   RebuildAppMenuRendererRequest,
 } from '../../../common/ipc/api';
 // IpcChannel<Incoming, Outgoing>
-export const rebuildApplicationMenu: RendererIpcChannel<
+export type RebuildApplicationMenu = RendererIpcChannel<
   RebuildAppMenuMainResponse,
   RebuildAppMenuRendererRequest
-> = new RendererIpcChannel(REBUILD_APP_MENU_CHANNEL);
+>;
+export const rebuildApplicationMenu: RebuildApplicationMenu = new RendererIpcChannel(
+  REBUILD_APP_MENU_CHANNEL
+);